+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Precise UI Changelog
 
+## 1.5.0
+
+- Add `closeable` property for `Notification` component
+- Add `disableMobileFullWidth` property for `Button` component
+
 ## 1.4.2
 
 - Fix button documentation